WindowHandler: Implement the fundementals of tree based gui drawing
Updated the InterfaceElement with unique global ids so that we can use
it to find an element within a tree and place elements inside each other
such that they have relative posistioning. This implements the following
functions for InterfaceElement: getElementId, getChildElements,
addChildElement.
This also adds the virtual method getPosition which gets the relative
offsets of the element.

Added the virtual method getGlobalPos to the Drawable virtual class.
This gets the posittion of an element relative to the root window or the
postition (0, 0).

Implentented the getPosition and getGlobalPos in the window class.

Note: there is a bug with the realtive postitioning in the window class
constructor where we don't take into account that the give position is
ment to be an offset of the parent.